minor fix and clean up code
